fix: failing to find headers when linked with Pods. (#75)
Summary:
---------

Multiple people have reported that RNCAsyncStorage fails to build after #35. See https://github.com/react-native-community/react-native-async-storage/pull/35#issuecomment-482604705, #73, and #74.

After a little discussion, we agreed to rename the `.podspec` and the name of the Pod to be more in line with Apple's naming guidelines and the rest of the community. Unfortunately, this means that the next upgrade will break.

I got a little impatient knowing that I broke the latest release so I went ahead and implemented the fix. Sorry if you've already started looking at this.

Migration Plan:
---------------

Even though `react-native link` will be able to link the new `.podspec`, it won't clean up existing entries. Please find your Podfile, and make the following changes:

```diff
- pod 'react-native-async-storage', :path => '../node_modules/@react-native-community/async-storage'
+ pod 'RNCAsyncStorage', :path => '../node_modules/@react-native-community/async-storage'
```