fix(iOS): Don't reset the storage when the manifest file cannot be read because of iOS data protection (#270)
This is to distinguish "manifest file does not exist or is corrupted" and "have no access to the manifest file yet" cases.

Starting with an empty dictionary is a reasonable fallback in case the manifest does not exist or cannot be parsed because of being corrupted or incompatible. However doing so might lead to unnecessary data loss when the app simply has no access to its documents yet due to iOS data protection (if enabled and when the app launches while the device is locked due to e.g. geofencing, silent notification, external accessory, etc).