fix(windows): errors reported from getAllKeys/clear (#424)
`getAllKeys` was reporting the correct results, but also an empty array
error. Unlike the multi* methods, these expect a single error, not an
array of errors.
The fix is to change the type sent by the windows implementation from an
array to a single object. That object is also just a null value, since
there are no errors being reported currently by the windows
implementation.